+++ compiler/core/j.ml
@@ -77,6 +77,13 @@ and property_map = (property_name * expression) list
 and length_object = Js_op.length_object
 and delim = External_arg_spec.delim = DNone | DStarJ | DNoQuotes | DBackQuotes
 
+and record_rest_field = {
+  record_rest_label: string;
+  record_rest_ident: ident option;
+}
+
+and param = Ident_param of ident
+
 and expression_desc =
   | Length of expression * length_object
   | Is_null_or_undefined of expression  (** where we use a trick [== null ] *)
@@ -132,7 +139,7 @@ and expression_desc =
   | Var of vident
   | Fun of {
       is_method: bool;
-      params: ident list;
+      params: param list;
       body: block;
       env: Js_fun_env.t;
       return_unit: bool;
@@ -165,6 +172,7 @@ and expression_desc =
   | Null
   | Await of expression
   | Spread of expression
+  | Record_rest of record_rest_field list * expression
 
 and for_ident_expression = expression
 (* pure*)
@@ -327,6 +335,8 @@ and deps_program = {
         finish_ident_expression;
         property_map;
         length_object;
+        record_rest_field;
+        param;
         (* for_ident; *)
         required_modules;
         case_clause;
@@ -337,3 +347,18 @@ FIXME: customize for each code generator
 for each code generator, we can provide a white-list
 so that we can achieve the optimal
 *)
+
+let record_rest_field_idents fields =
+  List.filter_map (fun {record_rest_ident} -> record_rest_ident) fields
+
+let param_idents = function
+  | Ident_param id -> [id]
+
+let params_idents params = List.concat_map param_idents params
+
+let params_as_idents params =
+  let rec aux acc = function
+    | [] -> Some (List.rev acc)
+    | Ident_param id :: rest -> aux (id :: acc) rest
+  in
+  aux [] params